Crate cranelift_faerie

Top-level lib.rs for cranelift_faerie.

Users of this module should not have to depend on faerie directly.

Structs

FaerieBackend
A FaerieBackend implements Backend and emits ".o" files using the faerie library.

FaerieBuilder
A builder for FaerieBackend.

FaerieProduct
This is the output of Module's finish function. It provides functions for writing out the object file to memory or a file.

Enums

Format
An object file format.

Constants

VERSION
Version number of this crate.